Extracted the Timer Input into another component
