+++ src/components/PanelInputs.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef } from "react"
 import { secondsToFormattedHMS } from "../helpers/timeConversion"
 
 // This next variables are used for debugging the timer
-const TIMER_INCREMENT = 1
+import InputTimer from "./InputTimer"
 const TIMER_INTERVAL_MS = 25 // TODO: Reset to 1000 ms
 
 const PanelInputs = ({ createNewEntry }) => {
@@ -15,71 +15,17 @@ const PanelInputs = ({ createNewEntry }) => {
 
     let starterTimestamp = useRef(0);
 
-    // Whenever the timer status is changed, it either creates an interval or clears it to keep track of the seconds
-    useEffect(() => {
-        console.log("Timer status: ", timerStatus)
-        let timerSetInterval = null;
-        if (timerStatus === "running") {
-            starterTimestamp.current = Date.now();
-            console.log(`The timer is running now, the starterTimestamp will be ${starterTimestamp.current}`)
-            timerSetInterval = setInterval(() => {
-                setSeconds(prevTime => prevTime + TIMER_INCREMENT)
-            }, TIMER_INTERVAL_MS);
-        } else
-            clearInterval(timerSetInterval)
-        return () => clearInterval(timerSetInterval);
-    }, [timerStatus])
-
-    const handleSubmitNewEntry = event => {
-        event.preventDefault()
-        // Stop the timer
-        setTimerStatus("stopped")
-        console.log("Stopped the timer")
-        // This is the real amount of ms in between the timer being started and stopped
-        let rawDuration = Date.now() - starterTimestamp.current
-        // However, given that we can modify the speed of the timer for testing purpouses we will take that
-        // into account and adjust the timestamps for this duration. In a normal setting this adjustment would be 1.0
-        let endTimestamp = starterTimestamp.current + (rawDuration * TIMER_INCREMENT * (1000 / TIMER_INTERVAL_MS))
-        // Submit the new entry
-        console.log(`The value of starterTimestamp is ${starterTimestamp.current}`)
-        console.log(`The value of the ending timestamp is ${endTimestamp}`)
-        console.log(`For a duration value of  ${rawDuration}`)
-        createNewEntry(newEntry, { start: starterTimestamp.current, end: endTimestamp })
-        // Reset the form
-        setNewEntry("")
-        setSeconds(0)
+    const handleNewEntrySubmitted = (entryName, interval) => {
+        // TODO
+        createNewEntry(entryName, interval)
     }
 
     return (
         <div>
-            {/* Timer Display */}
-            <p>Time elapsed: {secondsToFormattedHMS(seconds)}s</p>
-            {/* New Entry Input */}
-            <form onSubmit={handleSubmitNewEntry}>
-                <input id="newEntryInput"
-                    name="newEntryInput"
-                    type="text"
-                    value={newEntry}
-                    onChange={event => setNewEntry(event.target.value)}
-                    placeholder="Input what you're working on" />
-                <div>
-                    {timerStatus === "stopped" &&
-                        <button onClick={() => setTimerStatus("running")}
-                            // Only allow the timer to start when there is text in the input field
-                            disabled={newEntry.trim() === ""}>
-                            Start
-                        </button>}
-                    {timerStatus === "running" &&
-                        <input id="submitNewEntry"
-                            name="submitNewEntry"
-                            type="submit"
-                            value="Stop"
-                            // Only allow submissions when the timer is stopped and there is text in the input field
-                            disabled={newEntry.trim() === ""} />}
-                </div>
-            </form>
-        </div >
-    );
+                    <InputTimer
+                        handleSubmit={handleNewEntrySubmitted}
+                    />
+        </div>
+    )
 }
-
 export default PanelInputs
